requirements.txt
numpy
pillow
matplotlib
scikit-learn
tensorflow>=2.6
opencv-python
tqdm
